MOVEONLY: Move versionbits info out of versionbits.o
